Refactor move methods to base class
